pyproject.toml
[project]
name = "conan_barbarian"
version = "0.1.0"
requires-python = ">=3.9"
description = "utility to help building cpp_info data in conanfile scripts"
authors = [
    {name = "Marco Ceriani", email = "marco-ceriani@users.noreply.github.com"},
]
dependencies = [
    "pytest-timeout>=2.3.1",
]
readme = "README.md"
license = {text = "WTFPL"}

[project.scripts]
conan-barbarian = "conan_barbarian.libshelper:main_cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
pytest = [
    "pytest-cov>=5.0.0",
]
check = [
    "mypy>=1.10.0",
]
